+++ panov.ilya/P1/main.cpp
@@ -1,18 +1,19 @@
 #include <iostream>
 #include <exception>
-#include "numOfchar.hpp"
+#include "nchar.hpp"
 
 int main()
 {
-  panov::NumOfchar numOfchar;
+  panov::Nchar nchar;
   try
   {
+    std::cout << "first var" << "\n";
     std::cin.exceptions(std::istream::failbit);
     int value = 0;
     do
     {
       std::cin >> value;
-      numOfchar(value);
+      nchar(value);
     }
     while (value != 0);
   }
@@ -24,6 +25,31 @@ int main()
   catch (const std::logic_error& e)
   {
     std::cerr << e.what() << "\n";
+    return 1;
+  }
+  std::cout << nchar() << "\n";
+
+  panov::Nchar numOfchar;
+  try
+  {
+    std::cout << "second var" << "\n";
+    std::cin.exceptions(std::istream::failbit);
+    int value = 0;
+    do
+    {
+      std::cin >> value;
+      numOfchar(value);
+    }
+    while (value != 0);
+  }
+  catch (const std::istream::failure& i)
+  {
+    std::cerr << "OnlyNumberPLS" << "\n";
+    return 1;
+  }
+  catch (const std::logic_error& i)
+  {
+    std::cerr << i.what() << "\n";
     return 1;
   }
   std::cout << numOfchar() << "\n";
